1.remove InstanceConfig.groovy 2.move InstanceConfig.groovy into notused package. remove 3.remove TestDeploy as it is out dated 4.create "history" package
