+++ src/main/groovy/com/sankuai/srq/deploy/ProjectMetaManager.groovy
@@ -4,6 +4,8 @@ import groovy.json.JsonSlurper
 
 import org.slf4j.LoggerFactory
 
+import com.sankuai.srq.deploy.notused.ProjectMeta;
+
 class ProjectMetaManager {
 	def static final logger = LoggerFactory.getLogger("DeployEngine")
