+++ src/main/groovy/com/sankuai/srq/deploy/history/DeployHistoryManager.groovy
@@ -1,0 +1,5 @@
+package com.sankuai.srq.deploy.history
+
+class DeployHistoryManager {
+
+}
